requirements.txt
streamlit
anthropic
python-dotenv
requests
yfinance
pandas_ta